update: 	[*] fix uint64 build error on 386 arch
 [formerly 0a902f043aeb1bebb55863081576dd4fe9d967e2 [formerly 0c7063086f193c44706fa9268f2f7e07e566bc06] [formerly da3116d5cb2896ec9438ff8d4a43ab7149936f1c [formerly bd576a307442955a0b1334b6ee837214fb393702]] [formerly 09dd35ba11c4ad302a2a50372d27f4be1e5f48e1] [formerly a97d6b6a0dc8fe7a7a605e14f47cd3c29cf8e560 [formerly aa774096105929fed39d93e808e9f6ebd3c57f9d] [formerly a97d6b6a0dc8fe7a7a605e14f47cd3c29cf8e560 [formerly aa774096105929fed39d93e808e9f6ebd3c57f9d] [formerly db0924bb7edf5d5ede2f961ee221f62e8f750b9b [formerly b412390a49edb134e09d81d4aa3b018126ecb2a3]]]] [formerly ac999ee575867653f73f3a6fe1e17be06d6a9f3e] [formerly 62efaa242e7498d8c664bba80ecf7ff11e767b34 [formerly 36cee3d1a8dd37d87a7989d5aa0ce1fb40306d42] [formerly 9bd2120a81e9bc99f0f2bfcb007cf14ca0126817 [formerly eeee0b4b66d9302f092e1d22e1819d3ea92588ac] [formerly 9bd2120a81e9bc99f0f2bfcb007cf14ca0126817 [formerly eeee0b4b66d9302f092e1d22e1819d3ea92588ac] [formerly 9bd2120a81e9bc99f0f2bfcb007cf14ca0126817 [formerly eeee0b4b66d9302f092e1d22e1819d3ea92588ac] [formerly 15be3c1eebe455e992be5707f469aafd427a57b9 [formerly 9e447054d5f6393135124d3fcd87b154a85e016e]]]]]]]
 [formerly ac999ee575867653f73f3a6fe1e17be06d6a9f3e]




Former-commit-id: 59e203c3521bcb5ad2177f1291123877ec8ec4ba [formerly 3569b816a5f83272088173d0650eb0feb45c8827]
Former-commit-id: ef695651fc8589d082ddf1548b0ea29a9509505e [formerly 821bf2c06cd26fb0c720d795adbf2eca9ef67a32] [formerly 873c5812bd3d74bc518ff1c02cb3d1f5e6cb0e04 [formerly 626fbb2623e0a212f723ca2171a3f78beec6e757]]
Former-commit-id: 3a5f179b98e995cbe9307993a3954b1c9a785c42 [formerly 9274f9dec97bfe2da19f30ff05a88559e3e533cf]
Former-commit-id: 381240beb92097cce67118e36f7f6a7018b6131b

+++ io/io.go
@@ -205,7 +205,7 @@ func MakeMetric(name string, tags map[string]string, fields map[string]interface
 		switch v.(type) {
 		case uint64:
 			if v.(uint64) > uint64(math.MaxInt64) {
-				l.Warnf("on input `%s', filed %s, get uint64 %d > MaxInt64(%d), dropped", name, k, v.(uint64), math.MaxInt64)
+				l.Warnf("on input `%s', filed %s, get uint64 %d > MaxInt64(%d), dropped", name, k, v.(uint64), uint64(math.MaxInt64))
 				delete(fields, k)
 			} else { // convert uint64 -> int64
 				fields[k] = int64(v.(uint64))
